nlife-backend/requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.6
Pillow>=9.5.0
drf-yasg>=1.21.5
PyYAML>=6.0
uritemplate>=4.1.1
inflection>=0.5.1
packaging>=23.1
pytz>=2023.3
gunicorn>=21.2.0
whitenoise>=6.5.0
python-dotenv>=1.0.0
dj-database-url>=2.1.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
